[refactor] move codes to each cpp files -create ShellCommandItem classes
