+++ TestShell_Excutor/ShellCommand.h
@@ -1,6 +1,7 @@
 #pragma once
 #include "TS_function.h"
 
+
 class ShellCommandItem {
 public:
 	ShellCommandItem() {}
@@ -10,23 +11,7 @@ protected:
 	iTS_SSD* ssd;
 };
 
-class Write : public ShellCommandItem {
-public:
-	Write() {}
-	Write(iTS_SSD* ssd) : ssd { ssd } {}
-	bool execute(unsigned int num1, unsigned int num2) override;
-private:
-	iTS_SSD* ssd;
-};
 
-class FullWrite : public ShellCommandItem {
-public:
-	FullWrite() {}
-	FullWrite(iTS_SSD* ssd) : ssd{ ssd } {}
-	bool execute(unsigned int num1, unsigned int num2) override;
-private:
-	iTS_SSD* ssd;
-};
 class ShellCommand {
 public:
 	ShellCommand(iTS_SSD* ssd) : ssd{ ssd } {
