+++ TestShell_Excutor/ShellErase.cpp
@@ -1,0 +1,5 @@
+#include "ShellErase.h"
+bool Erase::execute(unsigned int num1, unsigned int num2)
+{
+	return true;
+}
