+++ TestShell_Excutor/ShellFlush.cpp
@@ -1,0 +1,5 @@
+#include "ShellFlush.h"
+bool Flush::execute(unsigned int num1, unsigned int num2)
+{
+	return true;
+}
