+++ TestShell_Excutor/ShellFullRead.cpp
@@ -1,0 +1,6 @@
+#include "ShellFullRead.h"
+bool FullRead::execute(unsigned int num1, unsigned int num2)
+{
+	return true;
+}
+
