+++ TestShell_Excutor/ShellRead.cpp
@@ -1,0 +1,5 @@
+#include "ShellRead.h"
+bool Read::execute(unsigned int num1, unsigned int num2)
+{
+	return true;
+}
