+++ TestShell_Excutor/ShellTestScenarios.cpp
@@ -1,1 +1,8 @@
 #include "ShellTestScenarios.h"
+
+bool TestScenario::execute(unsigned int num1, unsigned int num2)
+{
+	unsigned int scenarioNum = num1;
+	return true;
+}
+
